Split Versioning3 query tests
