+++ tests/versioning_3_test.go
@@ -40,7 +40,6 @@ import (
 	"go.temporal.io/server/common/testing/testvars"
 	"go.temporal.io/server/common/testing/updateutils"
 	"go.temporal.io/server/common/worker_versioning"
-	"go.temporal.io/server/service/matching"
 	"go.temporal.io/server/service/worker/workerdeployment"
 	"go.temporal.io/server/tests/testcore"
 	"google.golang.org/protobuf/types/known/durationpb"
@@ -60,41 +59,8 @@ func (s *Versioning3Suite) setupEnv(opts ...testcore.TestOption) *VersioningTest
 	return setupVersioning3Env(s.T(), opts...)
 }
 
-func setupVersioning3Env(t *testing.T, opts ...testcore.TestOption) *VersioningTestEnv {
-	opts = append([]testcore.TestOption{
-		testcore.WithDynamicConfig(dynamicconfig.MatchingDeploymentWorkflowVersion, int(versioning3DeploymentWorkflowVersion)),
-
-		// Make sure we don't hit the rate limiter in tests
-		testcore.WithDynamicConfig(dynamicconfig.FrontendGlobalNamespaceNamespaceReplicationInducingAPIsRPS, 1000),
-		testcore.WithDynamicConfig(dynamicconfig.FrontendMaxNamespaceNamespaceReplicationInducingAPIsBurstRatioPerInstance, 1),
-
-		// This is overridden for tests using runTestWithMatchingBehavior.
-		testcore.WithDynamicConfig(dynamicconfig.MatchingNumTaskqueueReadPartitions, 4),
-		testcore.WithDynamicConfig(dynamicconfig.MatchingNumTaskqueueWritePartitions, 4),
-
-		// Overriding the number of deployments that can be registered in a single namespace. Done only for this test suite
-		// since it creates a large number of unique deployments in the test suite's namespace.
-		testcore.WithDynamicConfig(dynamicconfig.MatchingMaxDeployments, 1000),
-
-		// Keep deployment versions short because worker-deployment system workflow IDs must fit into 255 characters (database constraint).
-		testcore.WithTestVars(func(tv *testvars.TestVars) *testvars.TestVars {
-			return tv.WithDeploymentSeries("v3").WithBuildID("b")
-		}),
-	}, opts...)
-
-	return newVersioningTestEnv(t, opts...)
-}
-
 func (s *Versioning3Suite) runTestWithMatchingBehavior(testFn func(*VersioningTestEnv, *Versioning3Suite), opts ...testcore.TestOption) {
-	for _, behavior := range testcore.AllMatchingBehaviors() {
-		s.Run(behavior.Name(), func(s *Versioning3Suite) {
-			envOpts := append([]testcore.TestOption{}, opts...)
-			envOpts = append(envOpts, behavior.Options()...)
-			env := s.setupEnv(envOpts...)
-			behavior.InjectHooks(env)
-			testFn(env, s)
-		})
-	}
+	runVersioning3TestWithMatchingBehavior(s, testFn, opts...)
 }
 
 func (s *Versioning3Suite) updateUtils() updateutils.UpdateUtils {
@@ -312,219 +278,6 @@ func (s *Versioning3Suite) testWorkflowWithPinnedOverride(env *VersioningTestEnv
 			return env.respondCompleteWorkflow(tv, vbUnpinned), nil
 		})
 	env.verifyWorkflowVersioning(s, tv, vbUnpinned, tv.Deployment(), tv.VersioningOverridePinned(), nil)
-}
-
-func (s *Versioning3Suite) TestQueryWithPinnedOverride_NoSticky() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testQueryWithPinnedOverride(env, false)
-	})
-}
-
-func (s *Versioning3Suite) TestQueryWithPinnedOverride_Sticky() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testQueryWithPinnedOverride(env, true)
-	})
-}
-
-func (s *Versioning3Suite) TestPinnedQuery_DrainedVersion_PollersAbsent() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testPinnedQueryDrainedVersion(env, false, false)
-	},
-		testcore.WithDynamicConfig(dynamicconfig.VersionDrainageStatusRefreshInterval, 1*time.Second),
-		testcore.WithDynamicConfig(dynamicconfig.VersionDrainageStatusVisibilityGracePeriod, 1*time.Second),
-		testcore.WithDynamicConfig(dynamicconfig.PollerHistoryTTL, 500*time.Millisecond),
-	)
-}
-
-func (s *Versioning3Suite) TestPinnedQuery_DrainedVersion_PollersPresent() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testPinnedQueryDrainedVersion(env, true, false)
-	},
-		testcore.WithDynamicConfig(dynamicconfig.VersionDrainageStatusRefreshInterval, 1*time.Second),
-		testcore.WithDynamicConfig(dynamicconfig.VersionDrainageStatusVisibilityGracePeriod, 1*time.Second),
-	)
-}
-
-func (s *Versioning3Suite) TestPinnedQuery_RollbackDrainedVersion() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testPinnedQueryDrainedVersion(env, true, true)
-	},
-		testcore.WithDynamicConfig(dynamicconfig.VersionDrainageStatusRefreshInterval, 1*time.Second),
-		testcore.WithDynamicConfig(dynamicconfig.VersionDrainageStatusVisibilityGracePeriod, 1*time.Second),
-	)
-}
-
-func (s *Versioning3Suite) testPinnedQueryDrainedVersion(env *VersioningTestEnv, pollersPresent bool, rollback bool) {
-	tv := env.Tv()
-
-	// create version v1 and make it current
-	idlePollerDone := make(chan struct{})
-	go func() {
-		env.idlePollWorkflow(s, tv, true, ver3MinPollTime, "should not have gotten any tasks since there are none")
-		close(idlePollerDone)
-	}()
-	env.setCurrentDeployment(s, tv)
-	env.WaitForChannel(idlePollerDone)
-
-	wftCompleted := make(chan struct{})
-	env.pollWftAndHandle(s, tv, false, wftCompleted,
-		func(task *workflowservice.PollWorkflowTaskQueueResponse) (*workflowservice.RespondWorkflowTaskCompletedRequest, error) {
-			s.NotNil(task)
-			return env.respondCompleteWorkflow(tv, vbPinned), nil
-		})
-
-	env.startWorkflow(s, tv, tv.VersioningOverridePinned())
-	env.WaitForChannel(wftCompleted)
-	env.verifyWorkflowVersioning(s, tv, vbPinned, tv.Deployment(), tv.VersioningOverridePinned(), nil)
-
-	// create version v2 and make it current which shall make v1 go from current -> draining/drained
-	idlePollerDone = make(chan struct{})
-	tv2 := tv.WithBuildIDNumber(2)
-	go func() {
-		env.idlePollWorkflow(s, tv2, true, ver3MinPollTime, "should not have gotten any tasks since there are none")
-		close(idlePollerDone)
-	}()
-	env.setCurrentDeployment(s, tv2)
-	env.WaitForChannel(idlePollerDone)
-
-	// wait for v1 to become drained
-	s.Await(func(s *Versioning3Suite) {
-		resp, err := env.FrontendClient().DescribeWorkerDeploymentVersion(s.Context(), &workflowservice.DescribeWorkerDeploymentVersionRequest{
-			Namespace: env.Namespace().String(),
-			Version:   tv.DeploymentVersionString(),
-		})
-		s.NoError(err)
-		s.Equal(enumspb.VERSION_DRAINAGE_STATUS_DRAINED, resp.GetWorkerDeploymentVersionInfo().GetDrainageInfo().GetStatus())
-	}, time.Second*10, time.Millisecond*1000)
-
-	if !pollersPresent {
-		// simulate the pollers going away, which should make the query fail as now the version is drained + has no pollers polling it
-		time.Sleep(1 * time.Second) //nolint:forbidigo
-		versionStr := tv.Deployment().GetBuildId()
-		if versioning3DeploymentWorkflowVersion < workerdeployment.AsyncSetCurrentAndRamping {
-			versionStr = worker_versioning.ExternalWorkerDeploymentVersionToString(worker_versioning.ExternalWorkerDeploymentVersionFromDeployment(tv.Deployment()))
-		}
-
-		_, err := env.queryWorkflow(s.Context(), tv)
-		s.Error(err)
-		s.ErrorContains(err, fmt.Sprintf(matching.ErrBlackholedQuery, versionStr, versionStr))
-	} else {
-		// since the version still has pollers, the query should succeed
-		env.pollAndQueryWorkflow(s, tv, false)
-	}
-
-	if rollback {
-		// ramp traffic back to v1 to remove it from drained state
-		env.setRampingDeployment(s, tv, 50, false)
-
-		// wait for v1 to become ramping
-		s.Await(func(s *Versioning3Suite) {
-			resp, err := env.FrontendClient().DescribeWorkerDeploymentVersion(s.Context(), &workflowservice.DescribeWorkerDeploymentVersionRequest{
-				Namespace: env.Namespace().String(),
-				Version:   tv.DeploymentVersionString(),
-			})
-			s.NoError(err)
-			s.Equal(enumspb.WORKER_DEPLOYMENT_VERSION_STATUS_RAMPING, resp.GetWorkerDeploymentVersionInfo().GetStatus())
-		}, time.Second*10, time.Millisecond*1000)
-
-		// the ramping status is propagated to the task queues
-		env.waitForDeploymentDataPropagation(s, tv, versionStatusRamping, false, tqTypeWf)
-
-		// the version has pollers and is ramping making the query succeed
-		env.pollAndQueryWorkflow(s, tv, false)
-
-	}
-}
-
-func (s *Versioning3Suite) testQueryWithPinnedOverride(env *VersioningTestEnv, sticky bool) {
-	tv := env.Tv()
-
-	if sticky {
-		env.warmUpSticky(s, tv)
-	}
-
-	wftCompleted := make(chan struct{})
-	env.pollWftAndHandle(s, tv, false, wftCompleted,
-		func(task *workflowservice.PollWorkflowTaskQueueResponse) (*workflowservice.RespondWorkflowTaskCompletedRequest, error) {
-			s.NotNil(task)
-			return env.respondEmptyWft(tv, sticky, vbUnpinned), nil
-		})
-
-	// Wait for the version to be present in the task queue. Version existence is required before it can be set as an override.
-	s.Await(func(s *Versioning3Suite) {
-		resp, err := env.GetTestCluster().MatchingClient().CheckTaskQueueVersionMembership(s.Context(), &matchingservice.CheckTaskQueueVersionMembershipRequest{
-			NamespaceId:   env.NamespaceID().String(),
-			TaskQueue:     tv.TaskQueue().GetName(),
-			TaskQueueType: tqTypeWf,
-			Version:       worker_versioning.DeploymentVersionFromDeployment(tv.Deployment()),
-		})
-		s.NoError(err)
-		s.True(resp.GetIsMember())
-	}, 10*time.Second, 500*time.Millisecond)
-
-	runID := env.startWorkflow(s, tv, tv.VersioningOverridePinned())
-
-	env.WaitForChannel(wftCompleted)
-	env.verifyWorkflowVersioning(s, tv, vbUnpinned, tv.Deployment(), tv.VersioningOverridePinned(), nil)
-	if sticky {
-		env.verifyWorkflowStickyQueue(s, tv.WithRunID(runID))
-	}
-
-	env.pollAndQueryWorkflow(s, tv, sticky)
-}
-
-func (s *Versioning3Suite) TestUnpinnedQuery_NoSticky() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testUnpinnedQuery(env, false)
-	})
-}
-
-func (s *Versioning3Suite) TestUnpinnedQuery_Sticky() {
-	s.runTestWithMatchingBehavior(func(env *VersioningTestEnv, s *Versioning3Suite) {
-		s.testUnpinnedQuery(env, true)
-	})
-}
-
-func (s *Versioning3Suite) testUnpinnedQuery(env *VersioningTestEnv, sticky bool) {
-	tv := env.Tv()
-	tv2 := tv.WithBuildIDNumber(2)
-	if sticky {
-		env.warmUpSticky(s, tv)
-	}
-
-	wftCompleted := make(chan struct{})
-	env.pollWftAndHandle(s, tv, false, wftCompleted,
-		func(task *workflowservice.PollWorkflowTaskQueueResponse) (*workflowservice.RespondWorkflowTaskCompletedRequest, error) {
-			s.NotNil(task)
-			env.verifyWorkflowVersioning(s, tv, vbUnspecified, nil, nil, tv.DeploymentVersionTransition())
-			return env.respondEmptyWft(tv, sticky, vbUnpinned), nil
-		})
-
-	env.setCurrentDeployment(s, tv)
-	env.waitForDeploymentDataPropagation(s, tv, versionStatusCurrent, false, tqTypeWf)
-
-	runID := env.startWorkflow(s, tv, nil)
-
-	env.WaitForChannel(wftCompleted)
-	env.verifyWorkflowVersioning(s, tv, vbUnpinned, tv.Deployment(), nil, nil)
-	if sticky {
-		env.verifyWorkflowStickyQueue(s, tv.WithRunID(runID))
-	}
-
-	pollerDone := make(chan struct{})
-	go func() {
-		env.idlePollWorkflow(s, tv2, true, 5*time.Second, "new deployment should not receive query")
-		close(pollerDone)
-	}()
-	env.pollAndQueryWorkflow(s, tv, sticky)
-	env.WaitForChannel(pollerDone) // wait for the idle poller to complete to not interfere with the next poller
-
-	env.setCurrentDeployment(s, tv2)
-	env.waitForDeploymentDataPropagation(s, tv2, versionStatusCurrent, false, tqTypeWf)
-
-	go env.idlePollWorkflow(s, tv, true, ver3MinPollTime, "old deployment should not receive query")
-	// Since the current deployment has changed, task will move to the normal queue (thus, sticky=false)
-	env.pollAndQueryWorkflow(s, tv2, false)
 }
 
 func (s *Versioning3Suite) TestPinnedWorkflowWithLateActivityPoller() {
